+++ CHANGELOG.md
@@ -51,6 +51,7 @@ And please only add new entries to the top of this list, right below the `# Unre
 - Added `Window::set_ime_allowed` supported on desktop platforms.
 - **Breaking:** IME input on desktop platforms won't be received unless it's explicitly allowed via `Window::set_ime_allowed` and new `WindowEvent::Ime` events are handled.
 - On macOS, `WindowEvent::Resized` is now emitted in `frameDidChange` instead of `windowDidResize`.
+- On macOS, add `WindowExtMacOS::set_background_color` and  `WindowBuilderExtMacOS::with_background_color`.
 
 # 0.26.1 (2022-01-05)